Mji/confidence thresholding (#21)
* do the confidence thresholding to identify not certain predictions

* only the owner can delete the file

* black, flake8, isort

* fix missing list id2label check

* reduced logs

* fixed the computation of confidence thresholding

+++ src/api/caidapp/views.py
@@ -100,12 +100,7 @@ def media_file_update(request, media_file_id):
     """Show and update media file."""
     mediafile = get_object_or_404(MediaFile, pk=media_file_id)
     if request.method == "POST":
-        form = MediaFileForm(
-            request.POST,
-            instance=mediafile
-            # request.FILES,
-            # owner=request.user
-        )
+        form = MediaFileForm(request.POST, instance=mediafile)
         if form.is_valid():
 
             # get uploaded archive
@@ -167,7 +162,6 @@ def model_form_upload(request):
         form = UploadedArchiveForm(
             request.POST,
             request.FILES,
-            # owner=request.user
         )
         if form.is_valid():
 
@@ -203,15 +197,22 @@ def model_form_upload(request):
 def delete_upload(request, uploadedarchive_id):
     """Delete uploaded file."""
     uploadedarchive = get_object_or_404(UploadedArchive, pk=uploadedarchive_id)
-    uploadedarchive.delete()
+    if uploadedarchive.owner == request.user:
+        uploadedarchive.delete()
+    else:
+        messages.error(request, "Only the owner can delete the file")
     return redirect("/caidapp/uploads")
 
 
+@login_required
 def delete_mediafile(request, mediafile_id):
     """Delete uploaded file."""
     obj = get_object_or_404(MediaFile, pk=mediafile_id)
     parent_id = obj.parent_id
-    obj.delete()
+    if obj.parent.owner == request.user:
+        obj.delete()
+    else:
+        messages.error(request, "Only the owner can delete the file")
     return redirect("caidapp:uploadedarchive_detail", uploadedarchive_id=parent_id)
